Move resize helpers to separate module
